perf: reduce allocations in Welch::windowed_segments
`Iterator::flat_map` just requires each closure to return an
`impl Iterator`, not a concrete struct, so we can avoid allocating the
inner buffers and instead just return an iterator of `Complex` formed
from the signal and the window.

This shows performance gains of almost 20% on the benchmark also added
in this commit.